requirements.txt
asgiref @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_cchykh5sgr/croots/recipe/asgiref_1661287542702/work
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cycler==0.11.0
Django @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_eeor5o9jj8/croots/recipe/django_1661292534510/work
django-cors-headers==3.14.0
djangorestframework==3.14.0
fonttools==4.29.1
kiwisolver==1.3.2
matplotlib==3.5.1
mpmath==1.2.1
mrjob==0.7.4
mysqlclient==2.1.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyMySQL @ file:///C:/ci/pymysql_1610464861530/work
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
scipy==1.8.0
six==1.16.0
sqlparse @ file:///C:/b/abs_83jn6owfpc/croot/sqlparse_1668793499314/work
sympy==1.10.1
typing_extensions==4.3.0
tzdata @ file:///tmp/build/80754af9/python-tzdata_1611693927741/work
wincertstore==0.2
